+++ VoiceCraft.Core/Interfaces/IAudioEffect.cs
@@ -7,6 +7,6 @@ namespace VoiceCraft.Core.Interfaces
     {
         EffectType EffectType { get; }
 
-        void Process(VoiceCraftEntity from, VoiceCraftEntity to, ulong effectBitmask, Span<float> data, int count);
+        void Process(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask, Span<float> data, int count);
     }
 }
